archive/Lianhun-Agent-Protocol/pyproject.toml
# Dependency Management
[tool.poetry]
name = "lianhun-agent"
version = "0.1.0"
description = "The 2026 AI Agent Standardization & Deconstruction Protocol"
authors = ["Yuanqi <yuanqi@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.0"
langgraph = "^0.0.1"
pydantic = "^2.6.0"
fastapi = "^0.109.0"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
